Fix compiler warnings
renderers/opengl/renderer/renderer.cpp:1232:20: warning: variable ‘newGLTexture’ set but not used [-Wunused-but-set-variable]
stlgeometryloader.cpp:116:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/plugins/geometryloaders/default/stlgeometryloader.cpp
@@ -113,7 +113,7 @@ bool StlGeometryLoader::loadBinary(QIODevice *ioDev)
     quint32 triangleCount;
     stream >> triangleCount;
 
-    if (ioDev->size() != headerSize + sizeof(quint32) + (triangleCount * 50))
+    if (quint64(ioDev->size()) != headerSize + sizeof(quint32) + (triangleCount * 50))
         return false;
 
     m_points.reserve(triangleCount * 3);
